+++ grype/db/v5/differ/test-fixtures/dbs/base/5/metadata.json
